ipc: presentation_posn should be uint64_t type
The presentation_posn used for dai rendered bytes readback to
host side should be uint64_t type, according to the driver side
definition.

+++ src/audio/dai.c
@@ -65,7 +65,7 @@ struct dai_data {
 
 	uint32_t dai_pos_blks;	/* position in bytes (nearest block) */
 
-	volatile uint32_t *dai_pos;
+	volatile uint64_t *dai_pos; /* host can read back this value without IPC */
 };
 
 /* this is called by DMA driver every time descriptor has completed */
